feat(gui): add service and remote mode support
