+++ easytier/Cargo.toml
@@ -189,7 +189,7 @@ sys-locale = "0.3"
 ringbuf = "0.4.5"
 async-ringbuf = "0.3.1"
 
-service-manager = { git = "https://github.com/chipsenkbeil/service-manager-rs.git", branch = "main" }
+service-manager = { git = "https://github.com/EasyTier/service-manager-rs.git", branch = "main" }
 
 zstd = { version = "0.13" }
